refactor list order logic

+++ app/src/main/java/com/example/tasks/data/Task.kt
@@ -11,5 +11,5 @@ data class Task(
     val createdAt: Date = Date(),
     val updatedAt: Date = Date(),
     val isPinned: Boolean = false,
-    val customSortOrder: Int = 0
+    val customSortOrder: Long = 0L
 )
